feat: Export PDF-attachments in addition
Refs: #289
