+++ src/main/java/ch/sbb/polarion/extension/pdf_exporter/util/WildcardUtils.java
@@ -1,9 +1,11 @@
 package ch.sbb.polarion.extension.pdf_exporter.util;
 
 import ch.sbb.polarion.extension.generic.regex.RegexMatcher;
+import lombok.experimental.UtilityClass;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
+@UtilityClass
 public class WildcardUtils {
 
     public static @NotNull String toRegex(@Nullable String wildcard) {
