+++ src/main/java/ch/sbb/polarion/extension/pdf_exporter/widgets/BulkPdfExportWidgetRenderer.java
@@ -40,15 +40,11 @@ public class BulkPdfExportWidgetRenderer extends AbstractWidgetRenderer {
         String sort = sortByParameter.asLuceneSortString();
         this.dataSet = dataSetParameter.getFor().sort(sort).revision(null);
         this.items = this.dataSet.items();
-        if (!items.isEmpty()) {
-            switch (items.iterator().next().getOldApi().getPrototype().getName()) {
-                case "Module": dataType = "Documents"; break;
-                case "RichPage": dataType = "Pages"; break;
-                case "TestRun": dataType = "Test Runs"; break;
-                default: dataType = "Unknown"; break;
-            }
-        } else {
-            dataType = "Unknown";
+        switch (dataSetParameter.prototype()) {
+            case Document: dataType = "Documents"; break;
+            case RichPage: dataType = "Pages"; break;
+            case TestRun: dataType = "Test Runs"; break;
+            default: dataType = "Unknown"; break;
         }
         this.columns = columnsParameter.fields();
