+++ src/main/resources/webapp/pdf-exporter/js/bulk-pdf-exporter.js
@@ -212,7 +212,7 @@ const BulkPdfExporter = {
             const documentId = currentItem.dataset["id"];
             if (documentType === ExportParams.DocumentType.TEST_RUN) {
                 this.exportParams["urlQueryParameters"] = { id: documentId };
-                this.downloadTestRunAttachments(this.exportParams.projectId, documentId, this.exportParams.revision, this.exportParams.attachmentsFilter);
+                ExportCommon.downloadTestRunAttachments(this.exportParams.projectId, documentId, this.exportParams.revision, this.exportParams.attachmentsFilter);
             } else {
                 this.exportParams["locationPath"] = `${currentItem.dataset["space"]}/${documentId}`;
             }
@@ -224,7 +224,7 @@ const BulkPdfExporter = {
                 BulkPdfExporter.finishedCount += 1;
                 BulkPdfExporter.updateState(BULK_EXPORT_IN_PROGRESS);
                 const downloadFileName = fileName || `${currentItem.dataset["space"] ? currentItem.dataset["space"] + "_" : ""}${documentId}.pdf`; // Fallback if file name wasn't received in response
-                this.downloadBlob(responseBody, downloadFileName);
+                ExportCommon.downloadBlob(responseBody, downloadFileName);
                 this.startNextItemExport();
             }, errorResponse => {
                 this.errors = true;
@@ -255,53 +255,6 @@ const BulkPdfExporter = {
         }
     },
 
-    downloadTestRunAttachments: function (projectId, testRunId, revision = null, filter = null) {
-        let url = `/polarion/pdf-exporter/rest/internal/projects/${projectId}/testruns/${testRunId}/attachments?`;
-        if (revision) url += `&revision=${revision}`;
-        if (filter) url += `&filter=${filter}`;
-
-        SbbCommon.callAsync({
-            method: "GET",
-            url: url,
-            responseType: "json",
-            onOk: (responseText, request) => {
-                for (const attachment of request.response) {
-                    this.downloadAttachmentContent(projectId, testRunId, attachment.id, revision);
-                }
-            },
-            onError: (status, errorMessage, request) => {
-                console.error('Error fetching attachments:', request.response);
-            }
-        });
-    },
-
-    downloadAttachmentContent: function (projectId, testRunId, attachmentId, revision = null) {
-        let url = `/polarion/pdf-exporter/rest/internal/projects/${projectId}/testruns/${testRunId}/attachments/${attachmentId}/content?`;
-        if (revision) url += `&revision=${revision}`;
-
-        SbbCommon.callAsync({
-            method: "GET",
-            url: url,
-            responseType: "blob",
-            onOk: (responseText, request) => {
-                this.downloadBlob(request.response, request.getResponseHeader("Filename"));
-            },
-            onError: (status, errorMessage, request) => {
-                console.error(`Error downloading attachment ${attachmentId}:`, request.response);
-            }
-        });
-    },
-
-    downloadBlob: function(blob, fileName) {
-        const objectURL = (window.URL ? window.URL : window.webkitURL).createObjectURL(blob);
-        const link = document.createElement("a");
-        link.href = objectURL;
-        link.download = fileName
-        link.target = "_blank";
-        link.click();
-        link.remove();
-        setTimeout(() => URL.revokeObjectURL(objectURL), 100);
-    }
 }
 
 BulkPdfExporter.init();
